split build command gated dependency tests
